Move records-found message and CSV button into ResultsList
