+++ src/components/QueryBuilder/QueryBuilder.css
@@ -5,7 +5,7 @@
   padding-bottom: 4px;
 }
 
-.SubmitRow {
+.ResultsSummary {
   display: flex;
   justify-content: space-between;
   font-size: 1.2em;
